Rename Erc20 to Token
